api/requirements.txt

# Python dependencies for Vercel
requests>=2.31.0
google-generativeai>=0.3.2
openpyxl>=3.1.2
pydantic>=2.5.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
